+++ src/Farmer/Arm/Web.fs
@@ -394,7 +394,7 @@ type Certificate =
 
 type SlotConfigName = 
     { SiteName : ResourceName
-      SlotSettingNames: List<string> }
+      SlotSettingNames: string Set }
     interface IArmResource with
         member this.ResourceId = config.resourceId(this.SiteName/"slotconfignames")
         member this.JsonModel =