+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 Release Notes
 =============
-## 1.6.22
+## vNext
 * WebApp/Functions: Support for deployment slot settings with `slot_setting` and `slot_settings`
 
 ## 1.6.21